src/main/java/com/greenfoxacademy/springwebapp/troop/controllers/TroopController.java
package com.greenfoxacademy.springwebapp.troop.controllers;

import com.greenfoxacademy.springwebapp.kingdom.models.KingdomEntity;
import com.greenfoxacademy.springwebapp.security.CustomUserDetails;
import com.greenfoxacademy.springwebapp.troop.models.dtos.TroopListResponseDto;
import com.greenfoxacademy.springwebapp.troop.services.TroopService;
import lombok.AllArgsConstructor;
import org.springframework.http.ResponseEntity;
import org.springframework.security.core.Authentication;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RestController;

@RestController
@AllArgsConstructor
public class TroopController {
  public static final String URI = "/kingdom/troops";

  private final TroopService troopService;

  @GetMapping(TroopController.URI)
  public ResponseEntity<TroopListResponseDto> getTroopsOfKingdom(Authentication auth) {
    KingdomEntity kingdom = ((CustomUserDetails) auth.getPrincipal()).getKingdom();
    TroopListResponseDto troops = troopService.troopsToListDTO(kingdom);
    return ResponseEntity.ok(troops);
  }
}